requirements.txt

gradio
pdfplumber
easyocr
pillow
python-docx
transformers
torch
sentence-transformers
numpy
sentencepiece
deep-translator
